frontend-app-api: move app node instance creation into instantiateAppNodeTree
