+++ packages/frontend-app-api/src/wiring/graph/types.ts
@@ -20,11 +20,6 @@ import {
   ExtensionDataRef,
 } from '@backstage/frontend-plugin-api';
 
-/** @internal */
-export type Mutable<T> = {
-  -readonly [P in keyof T]: T[P];
-};
-
 /**
  * The specification for this node in the app graph.
  * @public
